feat: fix prod width
