+++ components/MainView/MainView.tsx
@@ -1,10 +1,24 @@
+import SideNavMenu from '../SideNav/SideNavMenu/SideNavMenu'
+import { Grid } from '@mui/material'
 import { Container } from '@mui/system'
-import { FileUploader } from '../FileUploader/FileUploader'
 
-export default function MainView() {
+interface SideNavProps {
+	children: React.ReactNode
+}
+
+const MainView = ({ children }: SideNavProps) => {
 	return (
-		<Container className='pt-16'>
-			<FileUploader />
+		<Container className='occupyWholeScreen max-h-min w-screen p-0 max-w-full'>
+			<Grid container>
+				<Grid container item xs={3}>
+					<SideNavMenu />
+				</Grid>
+				<Grid item xs={9} className='bg-lightGrey'>
+					{children}
+				</Grid>
+			</Grid>
 		</Container>
 	)
 }
+
+export default MainView
