+++ package.json
@@ -1,38 +1,39 @@
 {
-  "name": "saie-web-sigaa",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.13.6",
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.33",
-    "@types/react-dom": "18.0.11",
-    "eslint-config-next": "13.2.4",
-    "file-saver": "^2.0.5",
-    "install": "^0.13.0",
-    "next": "13.2.4",
-    "npm": "^9.7.2",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-dropzone": "^14.2.3",
-    "recharts": "^2.7.2",
-    "typescript": "5.0.3"
-  },
-  "devDependencies": {
-    "@types/file-saver": "^2.0.5",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.43.0",
-    "eslint-plugin-unused-imports": "^2.0.0",
-    "postcss": "^8.4.24",
-    "tailwindcss": "^3.3.2"
-  }
+	"name": "saie-web-sigaa",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint",
+		"prod": "next build && next start"
+	},
+	"dependencies": {
+		"@emotion/react": "^11.11.1",
+		"@emotion/styled": "^11.11.0",
+		"@mui/icons-material": "^5.11.16",
+		"@mui/material": "^5.13.6",
+		"@types/node": "18.15.11",
+		"@types/react": "18.0.33",
+		"@types/react-dom": "18.0.11",
+		"eslint-config-next": "13.2.4",
+		"file-saver": "^2.0.5",
+		"install": "^0.13.0",
+		"next": "13.2.4",
+		"npm": "^9.7.2",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
+		"react-dropzone": "^14.2.3",
+		"recharts": "^2.7.2",
+		"typescript": "5.0.3"
+	},
+	"devDependencies": {
+		"@types/file-saver": "^2.0.5",
+		"autoprefixer": "^10.4.14",
+		"eslint": "^8.43.0",
+		"eslint-plugin-unused-imports": "^2.0.0",
+		"postcss": "^8.4.24",
+		"tailwindcss": "^3.3.2"
+	}
 }
