+++ pages/_app.tsx
@@ -4,14 +4,14 @@ import { Footer } from '@/components/Footer/Footer'
 import { Header } from '@/components/Header/Header'
 import { Grid, Box } from '@mui/material'
 import Head from 'next/head'
-import SideNav from '@/components/SideNav/SideNav'
+import MainView from '@/components/MainView/MainView'
 import { useEffect, useState } from 'react'
 import DisclaimerModal from '@/components/Disclaimer/Disclaimer'
 
 export default function App({ Component, pageProps }: AppProps) {
 	return (
 		<>
-			<Head>
+			<Head >
 				<title>SARUE</title>
 				<meta
 					name='description'
@@ -20,13 +20,16 @@ export default function App({ Component, pageProps }: AppProps) {
 				<meta name='viewport' content='width=device-width, initial-scale=1' />
 				<link rel='icon' href='/favicon.ico' />
 			</Head>
-			<main>
+			<main style={{ maxWidth: '100%' }} className='occupyWholeScreen m-0'>
 				<DisclaimerModal />
-				<Grid className='max-h-screen column'>
+				<Grid
+					sx={{ width: '100%' }}
+					className='occupyWholeScreen max-h-screen column m-0'
+				>
 					<Header />
-					<SideNav>
-						<Component {...pageProps} />
-					</SideNav>
+					<MainView>
+						<Component className='m-0' class='test123' {...pageProps} />
+					</MainView>
 					<Footer />
 				</Grid>
 			</main>
