+++ styles/globals.css
@@ -3,7 +3,6 @@
 @tailwind utilities;
 
 :root {
-	--max-width: 1100px;
 	--border-radius: 12px;
 	--font-mono: ui-monospace, Menlo, Monaco, 'Cascadia Mono', 'Segoe UI Mono',
 		'Roboto Mono', 'Oxygen Mono', 'Ubuntu Monospace', 'Source Code Pro',
@@ -112,6 +111,11 @@ body {
 a {
 	color: inherit;
 	text-decoration: none;
+}
+
+div.MuiContainer-root.MuiContainer-maxWidthLg.occupyWholeScreen {
+	max-width: 100% !important;
+	padding: 0;
 }
 
 @media (prefers-color-scheme: dark) {
